+++ src/library/inputs/sdlpointer.cpp
@@ -40,12 +40,12 @@ Uint32 SDL_GetMouseState(int *x, int *y)
 {
     DEBUGLOGCALL(LCF_SDL | LCF_MOUSE);
     if (x != NULL)
-        *x = ai.pointer_x;
+        *x = game_ai.pointer_x;
     if (y != NULL)
-        *y = ai.pointer_y;
+        *y = game_ai.pointer_y;
 
     /* Translating pointer mask to SDL pointer state */
-    return SingleInput::toSDL2PointerMask(ai.pointer_mask);
+    return SingleInput::toSDL2PointerMask(game_ai.pointer_mask);
 }
 
 Uint32 SDL_GetGlobalMouseState(int *x, int *y)
@@ -84,7 +84,7 @@ Uint32 SDL_GetRelativeMouseState(int *x, int *y)
     oldy = game_ai.pointer_y;
 
     /* Translating pointer mask to SDL pointer state */
-    return SingleInput::toSDL2PointerMask(ai.pointer_mask);
+    return SingleInput::toSDL2PointerMask(game_ai.pointer_mask);
 }
 
 void SDL_WarpMouseInWindow(SDL_Window * window, int x, int y)
@@ -101,7 +101,7 @@ void SDL_WarpMouseInWindow(SDL_Window * window, int x, int y)
     event2.motion.which = 0; // TODO: Mouse instance id. No idea what to put here...
 
     /* Build up mouse state */
-    event2.motion.state = SingleInput::toSDL2PointerMask(ai.pointer_mask);
+    event2.motion.state = SingleInput::toSDL2PointerMask(game_ai.pointer_mask);
     event2.motion.x = x;
     event2.motion.y = y;
     event2.motion.xrel = game_ai.pointer_x - x;
@@ -132,7 +132,7 @@ void SDL_WarpMouse(Uint16 x, Uint16 y)
     event1.motion.which = 0; // TODO: Mouse instance id. No idea what to put here...
 
     /* Build up mouse state */
-    event1.motion.state = SingleInput::toSDL1PointerMask(ai.pointer_mask);
+    event1.motion.state = SingleInput::toSDL1PointerMask(game_ai.pointer_mask);
     event1.motion.x = x;
     event1.motion.y = y;
     event1.motion.xrel = (Sint16)(game_ai.pointer_x - x);