+++ src/library/main.cpp
@@ -151,6 +151,7 @@ void __attribute__((constructor)) init(void)
     ai.emptyInputs();
     old_ai.emptyInputs();
     game_ai.emptyInputs();
+    old_game_ai.emptyInputs();
 
     /* Initialize timers. It uses the initial time set in the config object,
      * so they must be initialized after receiving it.